+++ public/CHANGELOG.md
@@ -2,7 +2,9 @@
 
 ## v27.0.0
 
-_Status: Upcoming_
+_Status: Released_
+
+[Package](https://www.npmjs.com/package/@muibook/components/v/27.0.0)
 
 ## Header [End]